+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduce ports to represent native processes and added support for external ports and encoded ports in external terms
 - Added `atomvm:get_creation/0`, equivalent to `erts_internal:get_creation/0`
 - Added menuconfig option for enabling USE_USB_SERIAL, eg. serial over USB for certain ESP32-S2 boards etc.
-- Partial support for `erlang:fun_info/2`
+- Partial support for `erlang:fun_info/2` and `erlang:fun_info/1`
 - Added support for `registered_name` in `erlang:process_info/2` and `Process.info/2`
 - Added `net:gethostname/0` on platforms with gethostname(3).
 - Added `socket:getopt/2`
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `init:get_argument/1`, `init:get_plain_arguments/0` and `init:notify_when_started/1`
 - Added `application:get_env/2`
 - Added CodeQL analysis to esp32, stm32, pico, and wasm workflows
+- Added Function.ex and Protocol.ex improving Elixir 1.18 support
 
 ### Changed